Implement Block, LocalVar, LocalConst and Assign translation logic

+++ src/ast/type_registry/mod.rs
@@ -48,14 +48,14 @@ impl TypeRegistry {
         self.path_to_id.get(type_path).copied()
     }
 
-    pub fn get_type_size(&self, type_id: &ResolvedType) -> usize {
+    pub fn get_type_size(&self, type_id: &ResolvedType) -> i32 {
         match type_id {
             ResolvedType::Primitive(ty) => ty.size(),
             ResolvedType::Struct(id) => self.structs[id].total_size,
         }
     }
 
-    pub fn get_type_alignment(&self, type_id: &ResolvedType) -> usize {
+    pub fn get_type_alignment(&self, type_id: &ResolvedType) -> i32 {
         match type_id {
             ResolvedType::Primitive(ty) => ty.alignment(),
             ResolvedType::Struct(id) => self.structs[id].alignment,
